Add attribute to explicitly disable cleanup on slave
This patch adds "skip_cleanup" attribute to the machineconfig's info tag.
If set to "1" or "true" the slave machine's preparation cleanup phase is
skipped and overrides the cleanup (-c) controller option.

Without this attribute it's impossible to run the controller and slave
instances on the same machine with virtual guests or 8021q/bonding/team
driver involved because the cleanup phase removes these network modules
before test execution and breaks the controller's network configuration.

Signed-off-by: Jan Tluka <jtluka@redhat.com>