Incorrect mode for file reading
Incorrect mode for file reading was specified for files to be copied to
slaves. We need only 'read,binary' mode.

Using current mode results in crash of the controller when run by
ordinary non-root user because such user does not have modify access rights
for the files in /usr/share.

Signed-off-by: Jan Tluka <jtluka@redhat.com>