TestPacketAssert: minor bug fixes
When the TestPacketAssert module was interrupted it didn't read the
remaining tcpdump output. Because of that the recieved packets count
would be less than what was really recieved.

This patch fixes that by making the variable line an attribute of the
class and adding the code for output handling to the interrupt handler.

There is also a minor change here, the "filter" option of the command is
no longer mandatory as it is not needed for correct execution of the
tcpdump command.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>