ctl: Recent XML format changes
This patch implements the XML format changes we discussed recently.

What used to be <machineconfig> is now called <slavemachine>:

    <slavemachine>
        <params>
            <param name="hostname" value="192.168.122.10"/>
            <param name="libvirt_domain" value="Fedora16"/>
            <param name="os"   value="fedora16"/>
            <param name="type" value="virtual"/>
        </params>

        <netdevices>
            <netdevice network="test-net" phys_id="1">
                <params>
                    <param name="type"   value="eth"/>
                    <param name="hwaddr" value="52:54:00:e4:75:16"/>
                    <param name="driver" value="virtio"/>
                </params>
            </netdevice>
        </netdevices>
    </slavemachine>

<machinerequires> tag in the recipe changed to <requirements>:

    <requirements>
        <params>
            <param name="os" value="rhel6.2"/>
        </params>
        <netdevices>
            <netdevice network="ttnet" phys_id="t1"/>
            <netdevice network="ttnet" phys_id="t2"/>
        </netdevices>
    </requirements>

Both machines and netdevices now have variable tag called <params>
that can be used to specify different parameters and also to match
accordingly to these parameters.

WARNING:

Changes in this patch are not yet done. Although the code is functional,
these changes will require larger refactoring of the code, especially
in the NetTestController class.

We expect some changes to the controller from Ondra in the following
days, so I decided to postpone the refactoring and do it afterwards.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>