XmlTemplates: Accessing $recipe no longer allowed
This commit removes the support for accessing the recipe dictionary
using the $recipe template variable.

These are implementation details and we should not allow users to
access them directly.

Additionally, we will do a major refactoring of the NetTestController
class in the near future and this variable might even no longer exist
after that.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>