NetTestParse: Fix handler exceptions
The parser wraps exceptions that go through his handler hooks to his own
exception class called XmlProcessingError to add information about
at which line of the XML file the problem happened.

Doing this however results in discarding the original traceback, which
is replaced by different one that leads to the try/catch block. This
is not helpful at all. It actually makes debugging extremelly irritating.

This patch changes this behaviour so the parser only logs an error with
the line information and it re-raises the original exeption with the
original traceback.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>