Logs: code removal
This commit removes old deprecated code, in preparation for the
reimplementation of how we work with the logging module.

I completely removed the class stdLogger because its sole purpose was to
introduce a side-effect when importing the file. We decided that we
don't want this functionality or any side-effects anymore.

I also removed the static class Logs because the code was too complex to
use, understand and maintain. The reimplementation is focusing on this
class, the new implementation will be present in the following commit.

Finally one smaller change- I removed the global constant LOCAL_IP and
placed its value into a variable of the MultilineFormatter class as this
was the only place it was ever used and I failed to see a reason to keep
it as a global constant.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>