Logs: add class LoggingCtl
This commit adds a new class LoggingCtl. This class replaces the
functionality of the old static class Logs. This class however is not
static. Instead we will create one instance of this class that will
carry all the important state information needed to properly manipulate
with the logging module.

The class implements a simple interface that reflects our needs when
running test recipes:

* {set,unset}_recipe- when we loop over a list of specified recipes
* {add,remove}_slave- when connecting to new slaves
* {set,unset}_connection- for the slave application to redirect where to
                          send log records

When used correctly the class will output logs on the display and can
create a directory structure like this:

<log_folder>/                        (for summary of recipe results)

<log_folder>/<recipe>/               (for all the recipe specific logs)

<log_folder>/<recipe>/<slave>/       (for slave specific logs)

This patch also changes the output format of log messages to what we
have agreed on.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>