NetTestCommand: add pipe, set logs
This commit adds a new pipe connection for background commands. This
pipe will be used to transmit log messages and later it will overtake
the entire communication with the background process.

I also added a method to the NetTestCommandContext class that will
retrieve the communication pipes from all the background processes- this
is a temporary solution to allow NetTestSlave to register these pipes in
its ServerHandler.

I also made the background processes to correctly use these pipes to
transport logs to the Slave application which will realy them to the
controller.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>