NetTestSlave: new Slave implementation
This commit adds a new class ServerHandler, reimplements the
NetTestSlave class and renames the class NetTestSlaveXMLRPC to
SlaveMethods.

The NetTestSlave class implements the control flow of the slave
machines. Its run method contains the main loop in which we try to
accept a new connection and proccess recieved messages. The messages are
processed by the function _process_msg and if the message represents a
command recieved from the controller it is relayed to a SlaveMethods
object.

The handling of sockets and pipes used in communication is located in
the class ServerHandler that contains the server socket and the socket
representing the connection from the controller. And to track the
controller socket and the communication pipes it uses a
ConnectionHandler object.

I renamed the NetTestSlaveXMLRPC class to SlaveMethods because it more
closely resebles how it is used by the NetTestSlave.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>