NetTestController: reimplement communication
This commit reimplements the communication from the controllers
perspective.

I added a new class MessageDispatcher that inherits the class
ConnectionHandler. This class adds new methods that can be easily used
to send messages to the slaves. It encapsulates communication with every
slave used by the controller. For this reason this patch also removes
the rpc objects from machines infos and related methods. Instead we
create a normal TCP socket that is stored in the MessageDispatcher
object.

For compatibility reasons I simple rewrote the _rpc_call method so that
the communication just works with through the MessageDispatcher but I
expect that this will change in the future when we refactor the rest of
the controller.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>