recipes: Adding Smoke test recipes
We are breaking quite a few things at the moment (and we plan to keep
doing so for a while), so I decided to develop a small set of smoke
tests we can use while working on these changes.

These tests should cover most of the basic functionality and verify
that the changes didn't break anything crucial.

If you have any more things that should be tested, please, feel free
to add them here!

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>