NetTestCommand: reimplementation of NetTestCommand
This commit merges the classes BgCommand and NetTestCommand into one
single NetTestCommand class. The reason behind this is that we want all
test commands to be launched in the background because of how our new
ctl<->slave communication works and also for the cases where something
goes wrong in the test execution.

The new NetTestCommand class therefore handles the creation of the
forked processes and also the execution of the commands.

With this change in classes I also reimplemented our way of creating a
process in the background to use the multiprocessing module.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>