NetTestCommand: reimplement the NetTestCommand classes
These classes previously worked only with background processes and
needed to be changed to reflect this.

The kill/intr/wait test commands are the only exception to the
"all test commands will run in the background" and will be run in the
main process since they need access to the child processes.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>