NetTestSlave: adjust response handling to fit NetTestCommand
This patch reiplements the method run_command to reflect the changes of
NetTestCommand class. The method returns the result of the run() method
of the NetTestCommand class which is None for most cases but in case of
wait/intr/kill commands can be the result of the test command running in
background. And in the case of running a background command in the
traditional sense we still return a notification that yes the command
was forked.

Because of this the server also had to be adjusted a little. Currently
if the server recieves a return value from the run_command method it
will forward it to the controller. However if None is returned, nothing
is send to the controller which causes the controller to wait until a
response arrives. This way I simulate the wait for the end of a
foreground test command while also being able to recieve messages (such
as logs) from other slaves.

Now if the server recieves a result message this could mean 3 things:
* a foreground command finished and the server forwards the result to
  the controller

* a background command that wasn't killed/interrupted/waited on has
  finished- the server stores the result of this command and will
  forward it to the controller when the kill/intr/wait command is
  recieved

* a background command that was already killed/interrupted/waited on has
  finished- the server forwards the result to the controller.

This way the controller is waiting for a result message for every
command that he sends and at the same time he will be able to continue
when a background command has been launched.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>