NetTestSlave: add handling for exception messages
If an exception is raised in a command a message of type "exception" is
sent to the slave server. This commit adds support for correct handling
of such messages.

I also added an additional field to the exception message that can be
used to identify the command from which the message arrived. This is
used both in the log message created and to correctly remove the command
after it's failed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>