NetTestSlave: add kill_cmds method and cleanup
This commit adds a new slave method kill_cmds that will perform a
cleanup on the command context. This should be used by the controller
during slave cleanup phase, so that there are no leftover commands
running.

I also added a few calls of command context cleanup to the slave server
so that if a connection from controller is lost we also cleanup all
running commands.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>