NetTestController: fix slave cleanup
This commit moves the rpc_call of method "bye" so that it is the last
method called when communicating with the slave. It also adds a call of
the new "kill_cmds" method.

Finally I fixed the disconnect_slave method because it wasn't doing
anything usefull (the code was deleted somewhere in git history), now it
will properly disconnect the slaves.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>