SlavePool: test connection to slave machines
This commit adds a new functionality to the SlavePool class. Until now
we have used all configuration files located in the pool directory and
used them in our matching algorithm. This however won't work correctly
if some machines are turned off. In this case the matching algorithm
uses the first configuration that matches even if it contains a machine
that is turned off which will result in a failure state. Even though
there exists a match that is fully available.

This commit adds code to the SlavePool initialization where we test
connectability of each machine being added to the pool and skip machines
that are turned off.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>