NetTestCommand: fix deadlock bug
This commit fixes a deadlock situation that appears in case we interrupt
a command which results in the command ending a large amount of data at
once through the communication pipe. The pipe is buffered so the write
operation starts blocking until somebody starts reading from the pipe.
However the main process that would be reading this message is waiting
in a join() call creating a deadlock.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>