NetTestCommand: use the new disable_logging method
The NetTestCommand object only disabled the recipe handler and the
connection handler. This was incorrect as it ment that the logs from
commands wouldn't be saved on the slave machine. Also when there were
multiple commands running they would be accessing the tty without
synchronization which is probably not correct.

This commit changes the behaviour by calling the disable_logging method
which clears all the handlers to files, tty or sockets. And after
creates a new connection handler that forwards the log messages to the
main server process.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>