SlavePool: Support matching virtualized schemes
This commit adds native support for virtual machines with the
matching algorithm. Up to now, virtual machines could be only used
as if they were physical.

Now if the matching for physical machines fails, the controller
will try to find suitable set of virtual machines using which it could
create the required network.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>