NetTestController: handling of error messages
The communication protocol that was implemented can also send error
messages, however the Controller wasn't handling them and instead raised
an Unknown message exception. This patch fixes that by raising an
exception with the correct message.

I also modified the message that Slave sends when an unsupported method
call is recieved.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>