XML: Changing the 'interface' tag
This commit removes the <interface> tag from both the recipe and the
slavemachine config files. Now the interface tag's names are based
on their type.

For example an interface with type="bond" will now be represented by
a tag <bond ...>.

A slight cleanup of the slavemachine parser was done along with this
patch.