lnst-ctl: Adding match_setup action
This commit adds a new action called 'match_setup'. With this action,
the controller will only read the pool and try to match the recipe
setup to the pool. In case the match was successfull, it will print
the details of the match.

I didn't really pay that much attention to the format of the report.
That might change in the future. The important thing here are the
backend options for this functionality.