recipe format: <options> under <eth> moved under <slave>
Until now when setting options of team interface slaves the options had
to be specified under the <netdevice> or <eth> tag (in the new format).
However the new format didn't accept any <options> tags under the <eth>
tag. This commit fixes that by moving the <options> tag under its
respective <slave> tag. The team configuration now looks something like
this:

<eth id="1" network="A"/>
<team id="testiface">
    <options>
        <option>...</option>
    </options>
    <slaves>
        <slave id="1">
            <options>
                <option name="teamd_port_config">
                    {
                        "prio": -10,
                        "sticky": true
                    }
                </option>
            </options>
        </slave>
    </slaves>
    <addresses>
        <address>...</address>
    </addresses>
</team>

This doesn't change the behaviour of <options> tag that are under the
not-eth tags.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>