NetTestResultSerializer: add support for command_sequences
This commit extends the xml format used by the result serializer to
include command sequences of the original recipe. I added this because
we use a lot of recipes with multiple command sequences and it can get
pretty confusing trying to figure out which commands are in the same
command sequence, when you're not directly looking at the recipe.

I also added a 'result' attribute to the <recipe> node, that contains
the cumulative result of the recipe. This is a little bit redundant
because you can deduce it from the results of individual commands. But
it's more convenient like this.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>