NetTestResultSerializer: add method print_summary
This commit adds the method print_summary to the class
NetTestResultSerializer. The instances of this class contain recipe
results in DOM form. This was previously used to write results in XML
form to a file.

The test results can however be also used to provide a nice summary to
be printed at the end of lnst-ctl execution. This can now be done by
using the new method.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>