RecipeParse: restrict bg_id usage
The previous commit changed the execution of the system_config command,
which now cannot run in the background as it would break the
application. This commit adds a check to the recipe parser that will not
allow bg_id to be used in commands of type system_config, wait, intr and
kill.

This commit also updates the recipe schema accordingly.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>