NetTestResultSerializer: fix output of options
The serializer function we use transforms dict objects directly to xml.
For this it uses the keys as tag names. This however results in invalid
tags for system_config options which begin with a '/' character. This
commit fixes that by detecting if the parent tag is named 'options'.
This is not a very universal way to do this as these special characters
can theoretically appear anywhere, but a nice universal solution
probably doesnt exist without defining a strict result data interface,
which we currently don't have.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>