NetTestResultSerializer: fix recipe result on crash
In case of crash during recipe execution the result would be PASS even
though it should be FAIL. I tried to fix this before but it turns out I
only fixed it when the crash happened in the first command.

This commit should fix it completely, in the next patch I'll add an
error message to the command that caused the crash.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>