NetTestController: add command result on crash
If there was an exception raised during command execution, that command
wouldn't have a result assigned. This will be visible in the summary
output as a sequence of commands having passed but the overal recipe
result is FAIL.

This commit fixes that by adding a FAIL result to the command and an
error message to indicate that the command failed because of an
exception.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>