add attribute "pass_result" to <command> tags
The added attribute indicates what the expected result of the command
is. The result of the command is then calculated as an equality of the
expected result and recieved result. This can be used to test scenarios
where you expect the command should fail and therefore indicates a
successfull run of the command.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>