add generate-recipes.py
This commit adds a python version of the generate-recipes.sh script made
by Radek. I made this because I wanted to add a feature to the recipe
generation which required xml manipulation.

The new script will create a recipe for every combination of two
machines, similarily to how the original script did. After that it will
parse the recipe xml and expand the sourced tags. Then it will do a
final round of variable replacement. The resulting xml is written to the
recipe file.

The added feature is the second round of variable replacement. These
variables are dynamically loaded from the file variables.conf also added
in this patch. This file can define aditional variables in the section
defaults and then in sections named after the combinations of
configuration files (e.g. bond-bond or vlan-bond). This allows us to
define a slightly different behaviour of command sequences for different
configurations. At the moment the only defined variable is 'icmp_result'
which I use to indicate that the IcmpPing test should fail for
configurations of vlan-<non-vlan>.

To use a defined variable in the recipe fragments write this: #var_name#

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>