Colouring: Make the colour presets configurable
This patch adds support for a new set of configuration options to allow
the user to change the colour presets. The predefined colour presets are
the following:

    faded       for debugging output
    alert       for important messages
    highlight   highlighted text

    pass        PASS colour
    fail        FAIL colour

    info        the INFO label
    debug       the DEBUG label
    warning     the WARNING label
    error       the ERROR label
    log_header  the date and ip address in each log message

You can override these values on both the controller and the slave by
addint a [colours] section to your conf file and setting different
values for these presets. For example:

[colours]
faded = yellow default false
alert = red default true

Each colour preset comprises of three values:

    (foreground colour, background colour, bold flag)

You can use one of the 16 standard colours or (if your terminal supports
it) use the extended palette. The standard colours are:

    black, red, green, yellow, blue, magenta, cyan,
    light-gray, dark-gray, light-red, light-green,
    light-yellow, light-blue, light-magenta, light-cyan,
    white

The colours from the extended palette can be accessed by their index
as follows (256 colours are available):

    extended(0)
    extended(5)
    ...
    extended(255)

The complete palette can be found here:

    http://misc.flogisoft.com/bash/tip_colors_and_formatting

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>