Slave: add NmConfigDevice
This file contains a class hierarchy identical to the NetConfigDevice
class hierarchy. The difference is that NetworkManager is used for
device configuration.

This implementation is still very experimental so it is highly possible
that crashes will occur.

Currently macvlan and team are not supported by NM so these classes are
just placeholders for future implementation.

This class hierarchy will be used when NM is detected to be running on
the slave machine. The reason is that the normal tools won't work anyway
so we might as well try using NM.

Using a combination of NM and ip tools might be possible in future but
for now that would just complicate things.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>