Slave: use NmConfigDevice when NM is running
This patch adds detection of NM to the NetConfigDevice helper functions.
So when NM is running the classes in NmConfigDevice are used, instead of
those in NetConfigDevice.

I also changed the logged messages in NetTestSlave from error to warning
level, and reworded it so that it is obvious that the implementation is
highly experimental.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>