NetTestSlave: add kill_command method
The new method will kill the command with a specified id. The id is
either the bg_id of the command or None for foreground
commands, which makes the id a unique identifier. This method will be
used to forcefully kill a running command that has timed out.

This commit also modifies the message logged when killing a command.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>