NetTestSlave: method set_device_down detects NM
This method is automatically called by the controller when initializing
interfaces before they are configured.

When the interface is managed by NM calling ip link set <if> down will
make NM see it as unavailable to repair this we had to set the if back
to up before using NM. Instead when NM is present we will disconnect the
device through the dbus interface.

I didn't remove the code that "resets" the interface if it's unavailable
before using NM, just in case there will be similar problems in the
future.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>