Config: make lnst configuration globally accesible
LNST configuration was created as a single Config object that was passed
around as an argument to the modules that needed it. This was ok at
first because it was only needed in a few modules, but it became
tiresome to send this object everywhere as an argument.

This commit creates a single globally accesible object lnst_config that
is creted when the module lnst.Common.Config is first imported in
lnst-ctl or lnst-slave, where the contents of the object will be
initialized and loaded.

After that all the modules just import this object and can use it.

This commit also changes all the modules that were using the lnst
configuration so that they use this global object instead.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>