NetConfig: store references to NetConfigDevice objects
The NetConfig class created NetConfigDevice objects for one use only and
always discarded them. This works fine because we store the config
dictionaries that are needed to create these objects. However it
sometimes requires small hacks to work and it defeats the purpose of
using objects at all since they're not allowed to remember their state.

There will probably be some other changes made later to take advantage
of this.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>