NetTestSlave: cleanup on controller disconnect
This commit moves the construction of a NetConfig object to the
NetTestSlave class, so that it can call configuration cleanup when a
disconnect is detected. The disconnect handling is also changed to
reflect that.

I also made some minor code cleanup in this commit, but since it is all
related to the main part of this commit I didn't split it into a
sepparate commit.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>