NmConfigDevice: add NM management check functions
This commit adds a function is_nm_managed_by_name which checks if the
device with a specified name is managed by NM. When the device doesn't
exist it assumes that NM manages it as it's probably a soft device that
will be created.

Additionaly a class method is_nm_managed was added to the NmConfigDevice
class hierarchy that uses the new function and returns if the specified
device and potentially it's group of slaves is managed by NM and in case
of a conflict an Exception can be raised.

Lastly a function is_nm_managed has been added that just calls the
class method is_nm_managed from the approriate class.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>