NmConfigDevice: better detection of nm_managed
This commit adds functionality to the slave. Now when checking the
Managed property of a device if the device doesn't exist yet, we check
the slave devices first and if they are all not managed we declare that
the master device is also not managed. The same goes for all slaves
being managed.

This gives us a bit more flexibility when trying to mix NM configuration
with the traditional one.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>