ctl: Slightly changing the interface of SlavePool
This patch contains a small change in the provision_machines() function
that is used to find a match for a network setup. The dictionary for
results is no longer constructed within the object; it is pased to the
function via a parameter and the machines are filled in.

The method then returns a boolean value that indicates whether the
matching was successful or not.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>