ctl: Massive refactoring of the RecipeParser
Basicaly the whole parser was changed. The parser no longer directs the
execution of the controller. This means the controller hooks were
completely removed and all the semantic checks were moved to
NetTestController.

Apart from that, this commit also contains many structural changes in
the recipe XML file. The <commans_sequence> tag changed to <task>
and the all of the command were changed as well.

For the new format, please refer to the example recipes or to the
documentation that is available online.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>