ctl: Changing pass_result to expect
The pass_result was changed to expect. Additionaly, LNST was failing in
case the pass_result was false. In that case the "err_msg" was not set
properly. This patch also fixes the problem.

I added a FIXME comment there, because this (in my opinion) should be
done earlier on the slave rather than here. This way, the slave prints
ERROR as if the command failed (because it had in fact failed on the
slave), but the controller knows it was expected and puts PASS to the
summary. This is confusing.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>