NetTestResultSerializer: add error messages to recipe results
This commit extends result reporting of the run_recipe method in
NetTestController to report not only the result of the command but also
add an error message if the recipe failed. To reflect this
NetTestResultSerializer now uses that information to create an
additional xml node and the error message will appear in the result
summary.

The reason is that when the recipe failed during configuration there
would be no information in the result summary indicating what happend
only a that the recipe failed.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>