NetTestSlave: update of slave cleanup
I didn't notice this before but we already had a method that did slave
cleanup. This commit makes a slight modification to the function and
makes use of it instead of calling the individual cleanup methods
themselves.

The cleanup process is now logged in DEBUG level, and I added detection
of server startup so that no cleanup happens when we didn't do anything
yet.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>