NetTestSlave: listen for netlink notifications
The slave server now listens to NETLINK_ROUTE notifications. At the
moment only RTM_NEWLINK messages are interesting to us, so we construct
update messages from these and send them to the controller.

The update messages contain the name of the interface, it's hw address,
and the index under which it is recognized in the kernel. The index is
useless right now but it will be useful later when we will have an
internal interface database.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>