ConnectionHandler: handle NetlinkSockets
This commit adds the ability to recieve data from NetlinkSockets. The
data is also parsed and decoded by a NETLINK_ROUTE marshaller, since
that is the only type of netlink socket we will be using at the moment.

The order of the if/elif statements is important and the NetlinkSocket
must come before the socket.SocketType check.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>