Controller: add handling of update messages
This commit extends the message handling on the controller, to recieve
update messages from slave machines.

A new method was added to the Machine objects- interface_update that
will be called when an update message is recieved. At the moment it only
looks up the corresponding interface (by devname) and updates its hw
address. In the future this could be used to update ip addresses, and
devnames them selves (when the kernel indices are stored).

This change also required the addition of a dictionary to the message
dispatcher object that maps machine id to the machine object.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>