NetUtils: rewrite scan_netdevs function
The scan_netdev function was starting to be a bit unreliable sometimes
causing crashes. It also wasn't a very pretty solution of the
problem.

The new scan_netdev function connects to a netlink socket and requests a
dump of all interfaces on the system, and parses the result to return a
similar data structure to the previous implementation of the function.
Again it contains the addditional kernel index of the interface that
will be used in the future.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>