install: Adding schemes to the distribution
The schemes need to be installed with the controller so it can use them
during the runtime. This patch adds them to the setup.py script.

By default, they are stored in /usr/share/lnst/, but the path can be
configured in the controller config file using the 'resource_dir'
option.

Signed-off-by: Radek Pazdera <rpazdera@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>