NetTestCommand: command result interface change
This commit changes the way commands report their results. The
NetTestCommandGeneric class no longer defines a set_result function, and
instead set_pass or set_fail functions should be used. These functions
create the result object (a dictionary) that contains relevant fields.
The functions also check the expect property of the command potentially
inverting the result of the command, with a relevant message added.

Previously there was an option to supply a message when calling the
set_pass or set_fail method from a test module. This was removed and
instead these messages should be moved into the res_data object of the
test module.

The main reason for this was to move the checking of expect to the slave
and to also move res_data formatting to the slave.

Both set_ functions accept a parameter that contains result data of the
command. This should be a dictionary that can be serialized into a valid
xml document and a check function for this will be added later.

The supplied res_data is formatted using the function format_res_data
which can be overridden in the subclasses if the user wants to customize
the formatting.

The function _format_cmd_res_header is used to format the first line of
the command summary, which differs by command type. This shouldn't be
overridden when implementing a custom test module.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>