NetTestCommand: fix control commands
Control commands need to be changed because of the previous commit.
Without this change the result summary would contain the incorrect
result header line for commands that were interrupted or killed before
finishing but would display information correctly when the command
finished before the control command.

The changes made in this commit ensure that the header line is
reformatted when appropriate.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>