NetTestResultSerializer: new results reimplementation
This commit reimplements the result serializer so that it doesn't
automatically create a DOM tree out of the supplied results. Instead
results are stored in a nice dictionary hierarchy.

This changed combined with moving the res_data formatting done on slaves
makes it very easy to print the summary. And at the same time be able to
easily create an xml file.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>