NetTestCommand: add method check_res_data
This method checks validity of result data that will eventually be
serialized into an xml. We have agreed that we will only accept result
data in form of nested dictionaries, lists and stringable objects.

The function checks the structure of the data and uses regexes to check
if valid element names and character data can be created.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>