NetTestCommand: set_fail bug fix
This commit fixes the use of function set_fail in several places in the
NetTestCommand module. These changes were overlooked when the change in
result reporting happened, sorry for that...

For ExecCmd the message that was previously there was removed as we
agreed on the meeting. The stderr report log was moved into debug logs
as well.

Signed-off-by: Ondrej Lichtner <olichtne@redhat.com>
Signed-off-by: Jiri Pirko <jiri@resnulli.us>